Add files via upload
This is a website I made using react. Is a recipe website that uses a free online api to showcase a lot of different recipes from a lot of diferent cuisines(Chinese,thai , italian etc.). It shows you what ingredients you need and gives you instructions on how to make the dish.