Add files via upload
This is the first website I created around one year ago when I started learning js, html, and css. It is not even close to being finished but I played around with some cool logos and features and at the time that felt amazing. I was trying to make a personal website for my father :)